+++ iceoryx_hoofs/testing/include/iceoryx_hoofs/testing/watch_dog.hpp
@@ -30,12 +30,22 @@ using namespace iox::units::duration_literals;
 class Watchdog
 {
   public:
-    Watchdog(const iox::units::Duration& timeToWait) noexcept
+    explicit Watchdog(const iox::units::Duration& timeToWait) noexcept
         : m_timeToWait(timeToWait)
     {
     }
 
+    Watchdog(const Watchdog&) = delete;
+    Watchdog(Watchdog&&) = delete;
+    Watchdog& operator=(const Watchdog&) = delete;
+    Watchdog& operator=(Watchdog&&) = delete;
+
     ~Watchdog() noexcept
+    {
+        reset();
+    }
+
+    void reset() noexcept
     {
         if (m_watchdog.joinable())
         {
@@ -44,14 +54,25 @@ class Watchdog
         }
     }
 
-    void watchAndActOnFailure(std::function<void()> f) noexcept
+    void watchAndActOnFailure(const std::function<void()>& actionOnFailure = std::function<void()>()) noexcept
     {
+        reset();
+
         m_watchdog = std::thread([=] {
             m_watchdogSemaphore.timedWait(m_timeToWait)
                 .and_then([&](auto& result) {
                     if (result == iox::posix::SemaphoreWaitState::TIMEOUT)
                     {
-                        f();
+                        std::cerr << "Watchdog observed no reaction after " << m_timeToWait << ". Taking measures!"
+                                  << std::endl;
+                        if (actionOnFailure)
+                        {
+                            actionOnFailure();
+                        }
+                        else
+                        {
+                            std::terminate();
+                        }
                         EXPECT_TRUE(false);
                     }
                 })