+++ iceoryx_posh/include/iceoryx_posh/gateway/gateway_base.hpp
@@ -1,4 +1,5 @@
 // Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -59,6 +60,7 @@ class GatewayBase
   protected:
     // Needed for unit testing
     GatewayBase() noexcept = default;
+    capro::Interfaces getInterface() const noexcept;
 
   protected:
     popo::InterfacePort m_interfaceImpl{nullptr};