+++ iceoryx_posh/include/iceoryx_posh/iceoryx_posh_types.hpp
@@ -42,6 +42,10 @@ class PublisherPortUser;
 class SubscriberPortRouDi;
 class SubscriberPortUser;
 } // namespace popo
+namespace capro
+{
+class ServiceDescription;
+}
 
 using PublisherPortRouDiType = iox::popo::PublisherPortRouDi;
 using PublisherPortUserType = iox::popo::PublisherPortUser;
@@ -131,10 +135,11 @@ constexpr uint32_t APP_MESSAGE_SIZE = 512U;
 
 // Processes
 constexpr uint32_t MAX_PROCESS_NUMBER = 300U;
-/// Maximum number of instances of a given service, which can be found.
+/// Maximum number of services, which can be found.
 /// This limitation is coming due to the fixed capacity of the cxx::vector (This doesn't limit the offered number of
 /// instances)
-constexpr uint32_t MAX_NUMBER_OF_INSTANCES = 50U;
+/// @todo #415 increase number back to 50 once service registry is available via shared memory
+constexpr uint32_t MAX_NUMBER_OF_SERVICES = 10U;
 
 // Nodes
 constexpr uint32_t MAX_NODE_NUMBER = 1000U;
@@ -249,7 +254,7 @@ using TimePointNs_t = std::chrono::time_point<BaseClock_t, DurationNs_t>;
 
 namespace runtime
 {
-using InstanceContainer = iox::cxx::vector<capro::IdString_t, MAX_NUMBER_OF_INSTANCES>;
+using ServiceContainer = iox::cxx::vector<capro::ServiceDescription, MAX_NUMBER_OF_SERVICES>;
 using namespace units::duration_literals;
 constexpr units::Duration PROCESS_WAITING_FOR_ROUDI_TIMEOUT = 60_s;
 constexpr units::Duration PROCESS_KEEP_ALIVE_INTERVAL = 3 * roudi::DISCOVERY_INTERVAL;  // > DISCOVERY_INTERVAL