+++ iceoryx_posh/source/capro/service_description.cpp
@@ -120,11 +120,6 @@ ServiceDescription::ServiceDescription(const IdString_t& service,
 
 bool ServiceDescription::operator==(const ServiceDescription& rhs) const noexcept
 {
-    if (!isValid() || !rhs.isValid())
-    {
-        return false;
-    }
-
     if (m_serviceString != rhs.m_serviceString)
     {
         return false;