+++ iceoryx_posh/source/gateway/toml_gateway_config_parser.cpp
@@ -21,12 +21,6 @@
 #include <regex>
 
 iox::cxx::expected<iox::config::GatewayConfig, iox::config::TomlGatewayConfigParseError>
-iox::config::TomlGatewayConfigParser::parse() noexcept
-{
-    return iox::config::TomlGatewayConfigParser::parse(DEFAULT_CONFIG_FILE_PATH);
-}
-
-iox::cxx::expected<iox::config::GatewayConfig, iox::config::TomlGatewayConfigParseError>
 iox::config::TomlGatewayConfigParser::parse(const roudi::ConfigFilePathString_t& path) noexcept
 {
     iox::config::GatewayConfig config;